Refactor the configuration to make sure it is properly excluded from the SpringBootTest
