+++ gateway-service/src/test/java/org/zowe/apiml/config/error/check/MessageServiceConfiguration.java
@@ -1,0 +1,23 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+package org.zowe.apiml.config.error.check;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.zowe.apiml.message.core.MessageService;
+import org.zowe.apiml.message.yaml.YamlMessageService;
+
+@TestConfiguration
+public class MessageServiceConfiguration {
+    @Bean
+    public MessageService messageService() {
+    return new YamlMessageService("/gateway-messages.yml");
+}
+}
